Don't use \gitAuthorDate in the \gitMark
\gitAuthorDate is the date that the most recent commit was authored, not
the date of the tagged release, so it's a bit misleading to include this
date behind the release number.

This commit removes it for the time being, though I've sent an inquiry
to the author of gitinfo2 to see if it would be possible to have the
package provide access to the metadata of the tag, as well as the
metadata of the most recent commit.